\name{coef.ensemble}
\alias{coef.ensemble}
\title{Calculcate coefficients for all models in ensemble}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Returns raw, t-value, absolute t-value, and standardised coefficent values.
}
\usage{coef.ensemble(object, ...)}
\arguments{
\item{object}{ensemble of models}
\item{...}{}
}

\details{}

\examples{}
\keyword{regression}
