\name{coef.ensemble}
\alias{coef.ensemble}
\title{Calculcate coefficients for all models in ensemble.}
\usage{\method{coef}{ensemble} (object, ...)}

\description{
  Calculcate coefficients for all models in ensemble.
  Returns raw, t-value, absolute t-value, and standardised
  coefficent values.
}
\keyword{regression}
\arguments{
  \item{object}{ensemble of models}
  \item{...}{other arguments ignored}
}
