\name{findmodels}
\alias{findmodels}
\title{General ensemble of models from models in global workspace}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{findmodels(modeltype = "lm", dataset, pattern)}
\arguments{
\item{modeltype}{model type}
\item{dataset}{if specified, all models must use this dataset}
\item{pattern}{pattern of model object names to match}
}

\details{}

\examples{}
\keyword{regression}
