\name{findmodels}
\alias{findmodels}
\title{General ensemble of models from models in global workspace'...}
\usage{findmodels(modeltype="lm", dataset, pattern)}

\description{
  General ensemble of models from models in global
  workspace'
}
\keyword{regression}
\arguments{
  \item{modeltype}{model class}
  \item{dataset}{if specified, all models must use this dataset}
  \item{pattern}{pattern of model object names to match}
}
