\name{fitbest}
\alias{fitbest}
\title{Generate best linear models}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Use the leaps package to generate the best subsets.
}
\usage{fitbest(formula, data, nbest=10, ...)}
\arguments{
\item{formula}{model formula}
\item{data}{data frame}
\item{nbest}{number of subsets of each size to record}
\item{...}{other arguments passed to \code{\link[leaps]{regsubsets}}}
}

\details{}

\examples{}
\keyword{regression}
