% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ggobi.ensemble}
\alias{ggobi.ensemble}
\title{Load model ensemble into GGobi with appropriate edge structure}
\usage{
ggobi.ensemble(data, ...)
}
\arguments{
\item{data}{model ensemble object}

\item{...}{other arguments ignored}
}
\description{
Load model ensemble into GGobi with appropriate edge structure
}
\examples{
y <- swiss$Fertility
x <- swiss[, -1]
mods <- fitall(y, x, "lm")
\dontrun{
library(rggobi)
ggobi(mods, swiss)
}
}
\keyword{dynamic}
\keyword{regression}

