\name{lmboot}
\alias{lmboot}
\title{Bootstrap linear models}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Generate linear models by bootstrapping observations
}
\usage{lmboot(formula, data, n=100)}
\arguments{
\item{formula}{model formula}
\item{data}{data set}
\item{n}{}
}

\details{}

\examples{}
\keyword{regression}
