% Generated by roxygen2 (4.0.0): do not edit by hand
\name{new_ensemble}
\alias{new_ensemble}
\title{Create a new ensemble of models.}
\usage{
new_ensemble(models, data)
}
\arguments{
\item{models}{list of models}

\item{data}{associated data frame}
}
\description{
Create a new ensemble of models.
}
\keyword{internal}

