% Generated by roxygen2 (4.0.0): do not edit by hand
\name{residuals.ensemble}
\alias{residuals.ensemble}
\title{Calculate residuals for all models in ensemble.}
\usage{
\method{residuals}{ensemble}(object, ...)
}
\arguments{
\item{object}{ensemble of models}

\item{...}{other arguments ignored}
}
\value{
data.frame of class \code{resid_ensemble}
}
\description{
Calculate residuals for all models in ensemble.
}
\seealso{
\code{\link{summary.resid_ensemble}}
}
\keyword{regression}

