\name{stdcoef}
\alias{stdcoef}
\title{Calculcate standardised coefficients for a model}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Refits the model with data standardised to mean 0, standard deviation 1.
}
\usage{stdcoef(model)}
\arguments{
\item{model}{model to refit with standardised data}
}

\details{}

\examples{}
\keyword{internal}
