\name{summary.ensemble}
\alias{summary.ensemble}
\title{Summarise ensemble of models with model fit statistics}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Returns degrees of freedom, log likelihood, R-squared, AIC, BIC and adjusted R-squared.
}
\usage{summary.ensemble(object, ...)}
\arguments{
\item{object}{ensemble of models}
\item{...}{}
}

\details{}

\examples{}
\keyword{regression}
