\name{summary.resid_ensemble}
\alias{summary.resid_ensemble}
\title{Residual summary for model ensemble}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Summarise residuals from ensemble
}
\usage{summary.resid_ensemble(object, data=NULL, ...)}
\arguments{
\item{object}{model residuals from \code{\link{residuals.ensemble}}}
\item{data}{}
\item{...}{}
}

\details{}

\examples{}
\keyword{regression}
