\name{summary.variable_ensemble}
\alias{summary.variable_ensemble}
\title{Summarise variable ensemble}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Summary function of coef(ensemble)
}
\usage{summary.variable_ensemble(object, ...)}
\arguments{
\item{object}{}
\item{...}{}
}

\details{Provides variable level statistics.}

\examples{}
\keyword{regression}
