% Generated by roxygen2 (4.0.0): do not edit by hand
\name{summary.ensemble}
\alias{summary.ensemble}
\title{Returns degrees of freedom, log likelihood, R-squared, AIC, BIC and
adjusted R-squared.}
\usage{
\method{summary}{ensemble}(object, ...)
}
\arguments{
\item{object}{ensemble of models}

\item{...}{other arguments ignored}
}
\description{
Returns degrees of freedom, log likelihood, R-squared, AIC, BIC and
adjusted R-squared.
}
\keyword{regression}

