\name{summary.variable_ensemble}
\alias{summary.variable_ensemble}
\title{Summarise variable ensemble.}
\usage{\method{summary}{variable_ensemble} (object, ...)}

\description{
  Summarise variable ensemble.
}

\details{
  Provides variable level statistics.
}
\keyword{regression}
\arguments{
  \item{object}{ensemble of models}
  \item{...}{other arguments ignored}
}
