% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.r
\name{coef.ensemble}
\alias{coef.ensemble}
\title{Calculcate coefficients for all models in ensemble.
Returns raw, t-value, absolute t-value, and standardised coefficent values.}
\usage{
\method{coef}{ensemble}(object, ...)
}
\arguments{
\item{object}{ensemble of models}

\item{...}{other arguments ignored}
}
\description{
Calculcate coefficients for all models in ensemble.
Returns raw, t-value, absolute t-value, and standardised coefficent values.
}
\keyword{regression}
