% Generated by roxygen2 (4.0.0): do not edit by hand
\name{coef.ensemble}
\alias{coef.ensemble}
\title{Calculcate coefficients for all models in ensemble.
Returns raw, t-value, absolute t-value, and standardised coefficent values.}
\usage{
\method{coef}{ensemble}(object, ...)
}
\arguments{
\item{object}{ensemble of models}

\item{...}{other arguments ignored}
}
\description{
Calculcate coefficients for all models in ensemble.
Returns raw, t-value, absolute t-value, and standardised coefficent values.
}
\keyword{regression}

