\name{findmodels}
\alias{findmodels}
\title{Find models in workspace}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
General ensemble of models from models in global workspace
}
\usage{findmodels(modeltype = "lm", dataset, pattern)}
\arguments{
\item{modeltype}{model type}
\item{dataset}{if specified, all models must use this dataset}
\item{pattern}{pattern of model object names to match}
}

\details{}

\examples{}
\keyword{regression}
