% Generated by roxygen2 (4.0.0): do not edit by hand
\name{findmodels}
\alias{findmodels}
\title{General ensemble of models from models in global workspace'}
\usage{
findmodels(modeltype = "lm", dataset, pattern)
}
\arguments{
\item{modeltype}{model class}

\item{dataset}{if specified, all models must use this dataset}

\item{pattern}{pattern of model object names to match}
}
\description{
General ensemble of models from models in global workspace'
}
\keyword{regression}

