% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.r
\name{fitall}
\alias{fitall}
\title{Fit all combinations of x variables ($2^p$).}
\usage{
fitall(y, x, method = "lm", ...)
}
\arguments{
\item{y}{vector y values}

\item{x}{matrix of x values}

\item{method}{name of method used to fit the model, e.g
\code{\link{lm}},\code{\link[MASS]{rlm}}}

\item{...}{other arguments passed on to \code{method}}
}
\description{
This technique generalises \code{\link{fitbest}}.  While it is much
slower it will work for any type of model.
}
\examples{
y <- swiss$Fertility
x <- swiss[, -1]
mods <- fitall(y, x, "lm")
}
\keyword{regression}
