% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.r
\name{fitbest}
\alias{fitbest}
\title{Use the leaps package to generate the best subsets.}
\usage{
fitbest(formula, data, nbest = 10, ...)
}
\arguments{
\item{formula}{model formula}

\item{data}{data frame}

\item{nbest}{number of subsets of each size to record}

\item{...}{other arguments passed to \code{\link[leaps]{regsubsets}}}
}
\description{
Use the leaps package to generate the best subsets.
}
\examples{
y <- swiss$Fertility
mods <- fitbest(Fertility ~ ., swiss)
}
\keyword{regression}
