\name{ggobi.ensemble}
\alias{ggobi.ensemble}
\title{Explore a model ensemble with GGobi}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{ggobi.ensemble(data, original=NULL, ...)}
\arguments{
\item{data}{}
\item{original}{}
\item{...}{}
}

\details{}

\examples{}
\keyword{dynamic}
\keyword{regression}
