\name{lmboot}
\alias{lmboot}
\title{Generate linear models by bootstrapping observations}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{lmboot(formula, data, n=100)}
\arguments{
\item{formula}{model formula}
\item{data}{data set}
\item{n}{}
}

\details{}

\examples{}
\keyword{regression}
