\name{lmboot}
\alias{lmboot}
\title{Generate linear models by bootstrapping observations...}
\usage{lmboot(formula, data, n=100)}

\description{
  Generate linear models by bootstrapping observations
}
\keyword{regression}
\arguments{
  \item{formula}{model formula}
  \item{data}{data set}
  \item{n}{number of bootstrapped data sets to generate}
}
