% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lmboot}
\alias{lmboot}
\title{Generate linear models by bootstrapping observations}
\usage{
lmboot(formula, data, n = 100)
}
\arguments{
\item{formula}{model formula}

\item{data}{data set}

\item{n}{number of bootstrapped data sets to generate}
}
\description{
Generate linear models by bootstrapping observations
}
\keyword{regression}

