\name{residuals.ensemble}
\alias{residuals.ensemble}
\title{Calculate residuals for all models in ensemble}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{

}
\usage{residuals.ensemble(object, ...)}
\arguments{
\item{object}{ensemble of models}
\item{...}{}
}

\details{}
\seealso{\code{\link{summary.resid_ensemble}}}
\examples{}
\keyword{regression}
