\name{residuals.ensemble}
\alias{residuals.ensemble}
\title{Calculate residuals for all models in ensemble.}
\usage{\method{residuals}{ensemble} (object, ...)}

\description{
  Calculate residuals for all models in ensemble.
}
\value{data.frame of class \code{resid_ensemble}}
\seealso{\code{\link{summary.resid_ensemble}}}
\keyword{regression}
\arguments{
  \item{object}{ensemble of models}
  \item{...}{other arguments ignored}
}
