\name{summary.ensemble}
\alias{summary.ensemble}
\title{Returns degrees of freedom, log likelihood, R-squared, AIC, BIC and...}
\usage{\method{summary}{ensemble} (object, ...)}

\description{
  Returns degrees of freedom, log likelihood, R-squared,
  AIC, BIC and adjusted R-squared.
}
\keyword{regression}
\arguments{
  \item{object}{ensemble of models}
  \item{...}{other arguments ignored}
}
