% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.r
\name{summary.ensemble}
\alias{summary.ensemble}
\title{Returns degrees of freedom, log likelihood, R-squared, AIC, BIC and
adjusted R-squared.}
\usage{
\method{summary}{ensemble}(object, ...)
}
\arguments{
\item{object}{ensemble of models}

\item{...}{other arguments ignored}
}
\description{
Returns degrees of freedom, log likelihood, R-squared, AIC, BIC and
adjusted R-squared.
}
\keyword{regression}
