% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.r
\name{summary.resid_ensemble}
\alias{summary.resid_ensemble}
\title{Summarise residuals from ensemble.}
\usage{
\method{summary}{resid_ensemble}(object, data = attr(object, "data"), ...)
}
\arguments{
\item{object}{model residuals from \code{\link{residuals.ensemble}}}

\item{data}{associated data set}

\item{...}{other arguments ignored}
}
\description{
Summarise residuals from ensemble.
}
\keyword{regression}
