% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.r
\name{summary.variable_ensemble}
\alias{summary.variable_ensemble}
\title{Summarise variable ensemble.}
\usage{
\method{summary}{variable_ensemble}(object, ...)
}
\arguments{
\item{object}{ensemble of models}

\item{...}{other arguments ignored}
}
\description{
Provides variable level statistics.
}
\keyword{regression}
